refactoring: aggregator-main transaction mod.
