+++ aggregator/aggregator-main/Cargo.toml
@@ -9,17 +9,16 @@ edition = "2021"
 [dependencies]
 aggregator-common = { path = "../util/common", version = "0.1.0" }
 aggregator-rgbpp-tx = { path = "../util/rgbpp-tx", version = "0.1.0" }
-
 ckb-app-config = { path = "../../util/app-config", version = "= 0.116.1" }
 ckb-channel = { path = "../../util/channel", version = "= 0.116.1" }
 ckb-hash = { path = "../../util/hash", version = "= 0.116.1" }
 ckb-logger = { path = "../../util/logger", version = "= 0.116.1" }
 
-ckb-gen-types = { version = "=0.116.1", default-features = false }
-ckb-jsonrpc-types = "=0.116.1"
-ckb-types = "=0.116.1"
+ckb-sdk = { git = "https://github.com/ethanyuan/ckb-sdk-rust", branch = "v3.2.1-branch-chain", features = ["native-tls-vendored"] }
+ckb-gen-types = { package = "ckb-gen-types", git = "https://github.com/ethanyuan/ckb", branch = "v0.116.1-branch-chain" }
+ckb-jsonrpc-types = { package = "ckb-jsonrpc-types", git = "https://github.com/ethanyuan/ckb", branch = "v0.116.1-branch-chain" }
+ckb-types = { package = "ckb-types", git = "https://github.com/ethanyuan/ckb", branch = "v0.116.1-branch-chain" }
 
-ckb-sdk = "=3.2.0"
 crossbeam-channel = "0.5.1"
 jsonrpc-core = "18.0"
 molecule = { version = "0.7.5", default-features = false }
