+++ aggregator/aggregator-main/src/lib.rs
@@ -3,41 +3,24 @@
 pub(crate) mod schemas;
 pub(crate) mod transaction;
 
-use crate::schemas::leap::{CrossChainQueue, Request, Requests};
+use crate::schemas::leap::Request;
 
-use aggregator_common::{
-    error::Error,
-    utils::{encode_udt_amount, privkey::get_sighash_lock_args_from_privkey, QUEUE_TYPE},
-};
+use aggregator_common::{error::Error, utils::encode_udt_amount};
 use aggregator_rgbpp_tx::RgbppTxBuilder;
 use ckb_app_config::{AggregatorConfig, AssetConfig, LockConfig, ScriptConfig};
 use ckb_channel::Receiver;
-use ckb_logger::{error, info, warn};
-use ckb_sdk::{
-    rpc::ckb_indexer::{Cell, Order},
-    rpc::CkbRpcClient as RpcClient,
-    traits::LiveCell,
-    traits::{CellQueryOptions, MaturityOption, PrimaryScriptType, QueryOrder},
-};
-use ckb_types::H256;
+use ckb_logger::{error, info};
+use ckb_sdk::rpc::CkbRpcClient as RpcClient;
 use ckb_types::{
-    core::FeeRate,
-    core::ScriptHashType,
-    h256,
-    packed::{Byte32, CellDep, OutPoint, Script},
+    packed::{CellDep, OutPoint, Script},
     prelude::*,
+    H256,
 };
 
-use std::collections::{HashMap, HashSet};
+use std::collections::HashMap;
 use std::str::FromStr;
 use std::thread::{self, sleep};
 use std::time::Duration;
-
-const CKB_FEE_RATE_LIMIT: u64 = 5000;
-
-/// Sighash type hash
-pub const SIGHASH_TYPE_HASH: H256 =
-    h256!("0x9bd7e06f3ecf4be0f2fcd2188b23f1b9fcc88e5d4b65a8637b17723bbda3cce8");
 
 ///
 #[derive(Clone)]
@@ -46,7 +29,6 @@ pub struct Aggregator {
     poll_interval: Duration,
     rgbpp_rpc_client: RpcClient,
     branch_rpc_client: RpcClient,
-    rgbpp_scripts: HashMap<String, ScriptInfo>,
     branch_scripts: HashMap<String, ScriptInfo>,
     rgbpp_assets: HashMap<H256, AssetInfo>,
     rgbpp_locks: HashMap<H256, Script>,
@@ -67,13 +49,6 @@ struct AssetInfo {
     pub script_name: String,
 }
 
-#[allow(dead_code)]
-pub(crate) enum RequestType {
-    CkbToBranch = 1,
-    BranchToCkb = 2,
-    BranchToBranch = 3,
-}
-
 impl Aggregator {
     /// Create an Aggregator
     pub fn new(config: AggregatorConfig, poll_interval: Duration, branch_chain_id: String) -> Self {
@@ -84,15 +59,14 @@ impl Aggregator {
             config.rgbpp_uri.clone(),
             config.rgbpp_scripts.clone(),
             config.rgbpp_custodian_lock_key_path.clone(),
-            config.branch_chain_capacity_provider_key_path.clone(),
-            config.branch_chain_token_manager_lock_key_path.clone(),
+            config.rgbpp_queue_lock_key_path.clone(),
+            config.rgbpp_ckb_provider_key_path.clone(),
         );
         Aggregator {
             config: config.clone(),
             poll_interval,
             rgbpp_rpc_client,
             branch_rpc_client,
-            rgbpp_scripts: get_script_map(config.rgbpp_scripts),
             branch_scripts: get_script_map(config.branch_scripts),
             rgbpp_assets: get_asset_map(config.rgbpp_assets),
             rgbpp_locks: get_rgbpp_locks(config.rgbpp_asset_locks),
@@ -121,16 +95,23 @@ impl Aggregator {
             }
 
             // get queue data
-            let rgbpp_requests = poll_service.get_rgbpp_queue_requests();
+            let rgbpp_requests = poll_service.rgbpp_tx_builder.get_rgbpp_queue_requests();
             let (rgbpp_requests, queue_cell) = match rgbpp_requests {
-                Ok((rgbpp_requests, queue_cell)) => (rgbpp_requests, queue_cell),
+                Ok((rgbpp_requests, queue_cell)) => {
+                    let rgbpp_requests: Vec<_> = rgbpp_requests
+                        .into_iter()
+                        .map(|r| Request::new_unchecked(r.as_bytes()))
+                        .collect();
+                    let queue_cell = OutPoint::new_unchecked(queue_cell.as_bytes());
+                    (rgbpp_requests, queue_cell)
+                }
                 Err(e) => {
                     error!("get RGB++ queue data error: {}", e.to_string());
                     continue;
                 }
             };
 
-            let leap_tx = poll_service.create_leap_tx(rgbpp_requests.clone(), queue_cell.clone());
+            let leap_tx = poll_service.create_leap_tx(rgbpp_requests.clone(), queue_cell);
             let leap_tx = match leap_tx {
                 Ok(leap_tx) => leap_tx,
                 Err(e) => {
@@ -158,7 +139,7 @@ impl Aggregator {
                 };
                 match wait_for_tx_confirmation(
                     poll_service.rgbpp_rpc_client.clone(),
-                    update_queue_tx,
+                    H256(update_queue_tx.0),
                     Duration::from_secs(600),
                 ) {
                     Ok(()) => {}
@@ -172,138 +153,6 @@ impl Aggregator {
 
             thread::sleep(poll_interval);
         }
-    }
-
-    pub(crate) fn build_message_queue_cell_search_option(&self) -> Result<CellQueryOptions, Error> {
-        let message_queue_type = self.get_rgbpp_script(QUEUE_TYPE)?;
-        let (message_queue_lock_args, _) =
-            get_sighash_lock_args_from_privkey(self.config.rgbpp_queue_lock_key_path.clone())?;
-        let message_queue_lock = Script::new_builder()
-            .code_hash(SIGHASH_TYPE_HASH.pack())
-            .hash_type(ScriptHashType::Type.into())
-            .args(message_queue_lock_args.pack())
-            .build();
-
-        let cell_query_option = CellQueryOptions {
-            primary_script: message_queue_type,
-            primary_type: PrimaryScriptType::Type,
-            with_data: Some(true),
-            secondary_script: Some(message_queue_lock),
-            secondary_script_len_range: None,
-            data_len_range: None,
-            capacity_range: None,
-            block_range: None,
-            order: QueryOrder::Asc,
-            limit: Some(1),
-            maturity: MaturityOption::Mature,
-            min_total_capacity: 1,
-            script_search_mode: None,
-        };
-        Ok(cell_query_option)
-    }
-
-    fn get_rgbpp_queue_requests(&self) -> Result<(Vec<Request>, OutPoint), Error> {
-        let (queue_cell, queue_cell_data) = self.get_rgbpp_queue_cell()?;
-        if queue_cell_data.outbox().is_empty() {
-            info!("No requests in queue");
-            return Ok((vec![], OutPoint::default()));
-        }
-        let request_ids: Vec<Byte32> = queue_cell_data.outbox().into_iter().collect();
-
-        let queue_out_point = queue_cell.out_point.clone();
-        let (_, witness_input_type) =
-            self.get_tx_witness_input_type(queue_cell.out_point, self.rgbpp_rpc_client.clone())?;
-        let requests = Requests::from_slice(&witness_input_type.raw_data()).map_err(|e| {
-            Error::TransactionParseError(format!("get requests from witness error: {}", e))
-        })?;
-        info!("Found {} requests in witness", requests.len());
-
-        // check requests
-        let request_set: HashSet<Byte32> = requests
-            .clone()
-            .into_iter()
-            .map(|request| request.as_bytes().pack().calc_raw_data_hash())
-            .collect();
-        let all_ids_present = request_ids.iter().all(|id| request_set.contains(id));
-        if all_ids_present {
-            Ok((requests.into_iter().collect(), queue_out_point.into()))
-        } else {
-            Err(Error::QueueCellDataError(
-                "Request IDs in queue cell data do not match witness".to_string(),
-            ))
-        }
-    }
-
-    fn get_rgbpp_queue_cell(&self) -> Result<(Cell, CrossChainQueue), Error> {
-        info!("Scan RGB++ Message Queue ...");
-
-        let queue_cell_search_option = self.build_message_queue_cell_search_option()?;
-        let queue_cell = self
-            .rgbpp_rpc_client
-            .get_cells(queue_cell_search_option.into(), Order::Asc, 1.into(), None)
-            .map_err(|e| Error::LiveCellNotFound(e.to_string()))?;
-        if queue_cell.objects.len() != 1 {
-            return Err(Error::LiveCellNotFound(format!(
-                "Queue cell found: {}",
-                queue_cell.objects.len()
-            )));
-        }
-        info!("Found {} queue cell", queue_cell.objects.len());
-        let queue_cell = queue_cell.objects[0].clone();
-
-        let queue_live_cell: LiveCell = queue_cell.clone().into();
-        let queue_data = queue_live_cell.output_data;
-        let queue = CrossChainQueue::from_slice(&queue_data)
-            .map_err(|e| Error::QueueCellDataDecodeError(e.to_string()))?;
-
-        Ok((queue_cell, queue))
-    }
-
-    fn _get_branch_cell_dep(&self, script_name: &str) -> Result<CellDep, Error> {
-        self.branch_scripts
-            .get(script_name)
-            .map(|script_info| script_info.cell_dep.clone())
-            .ok_or_else(|| Error::MissingScriptInfo(script_name.to_string()))
-    }
-
-    fn get_rgbpp_script(&self, script_name: &str) -> Result<Script, Error> {
-        self.rgbpp_scripts
-            .get(script_name)
-            .map(|script_info| script_info.script.clone())
-            .ok_or_else(|| Error::MissingScriptInfo(script_name.to_string()))
-    }
-
-    fn _get_branch_script(&self, script_name: &str) -> Result<Script, Error> {
-        self.branch_scripts
-            .get(script_name)
-            .map(|script_info| script_info.script.clone())
-            .ok_or_else(|| Error::MissingScriptInfo(script_name.to_string()))
-    }
-
-    fn fee_rate(&self) -> Result<u64, Error> {
-        let value = {
-            let dynamic = self
-                .rgbpp_rpc_client
-                .get_fee_rate_statistics(None)
-                .map_err(|e| Error::RpcError(format!("get dynamic fee rate error: {}", e)))?
-                .ok_or_else(|| Error::RpcError("get dynamic fee rate error: None".to_string()))
-                .map(|resp| resp.median)
-                .map(Into::into)
-                .map_err(|e| Error::RpcError(format!("get dynamic fee rate error: {}", e)))?;
-            info!("CKB fee rate: {} (dynamic)", FeeRate(dynamic));
-            if dynamic > CKB_FEE_RATE_LIMIT {
-                warn!(
-                    "dynamic CKB fee rate {} is too large, it seems unreasonable;\
-                so the upper limit {} will be used",
-                    FeeRate(dynamic),
-                    FeeRate(CKB_FEE_RATE_LIMIT)
-                );
-                CKB_FEE_RATE_LIMIT
-            } else {
-                dynamic
-            }
-        };
-        Ok(value)
     }
 }
