+++ aggregator/aggregator-main/src/transaction/mod.rs
@@ -1,1 +1,58 @@
 mod leap;
+
+use crate::Aggregator;
+
+use aggregator_common::error::Error;
+use ckb_logger::{info, warn};
+use ckb_types::{
+    core::FeeRate,
+    h256,
+    packed::{CellDep, Script},
+    H256,
+};
+
+pub const SIGHASH_TYPE_HASH: H256 =
+    h256!("0x9bd7e06f3ecf4be0f2fcd2188b23f1b9fcc88e5d4b65a8637b17723bbda3cce8");
+const CKB_FEE_RATE_LIMIT: u64 = 5000;
+
+impl Aggregator {
+    fn fee_rate(&self) -> Result<u64, Error> {
+        let value = {
+            let dynamic = self
+                .rgbpp_rpc_client
+                .get_fee_rate_statistics(None)
+                .map_err(|e| Error::RpcError(format!("get dynamic fee rate error: {}", e)))?
+                .ok_or_else(|| Error::RpcError("get dynamic fee rate error: None".to_string()))
+                .map(|resp| resp.median)
+                .map(Into::into)
+                .map_err(|e| Error::RpcError(format!("get dynamic fee rate error: {}", e)))?;
+            info!("CKB fee rate: {} (dynamic)", FeeRate(dynamic));
+            if dynamic > CKB_FEE_RATE_LIMIT {
+                warn!(
+                    "dynamic CKB fee rate {} is too large, it seems unreasonable;\
+            so the upper limit {} will be used",
+                    FeeRate(dynamic),
+                    FeeRate(CKB_FEE_RATE_LIMIT)
+                );
+                CKB_FEE_RATE_LIMIT
+            } else {
+                dynamic
+            }
+        };
+        Ok(value)
+    }
+
+    fn get_branch_cell_dep(&self, script_name: &str) -> Result<CellDep, Error> {
+        self.branch_scripts
+            .get(script_name)
+            .map(|script_info| script_info.cell_dep.clone())
+            .ok_or_else(|| Error::MissingScriptInfo(script_name.to_string()))
+    }
+
+    fn _get_branch_script(&self, script_name: &str) -> Result<Script, Error> {
+        self.branch_scripts
+            .get(script_name)
+            .map(|script_info| script_info.script.clone())
+            .ok_or_else(|| Error::MissingScriptInfo(script_name.to_string()))
+    }
+}
