--- aggregator/util/branch-tx/src/lib.rs
@@ -1,1 +1,0 @@
-#![allow(missing_docs)]
