+++ aggregator/util/rgbpp-tx/Cargo.toml
@@ -10,12 +10,11 @@ aggregator-common = { path = "../common", version = "0.1.0" }
 ckb-app-config = { path = "../../../util/app-config", version = "= 0.116.1" }
 ckb-stop-handler = { path = "../../../util/stop-handler", version = "= 0.116.1" }
 
+ckb-sdk = "=3.2.0"
 ckb-gen-types = { version = "=0.116.1", default-features = false }
 ckb-logger = "=0.116.1"
 ckb-jsonrpc-types = "=0.116.1"
 ckb-types = "=0.116.1"
-
-ckb-sdk = "=3.2.0"
 
 molecule = { version = "0.7.5", default-features = false }
 serde = { version = "1.0", features = ["derive"] }
